Align QR surface with updated API contract
- PmcQrMatrix: fields -> get-only properties, Ecc -> int, add
  this[x,y] indexer (glue consumes it), keep GetModule/ToRows/Modules.
- PmcQr.Encode now returns PmcQrMatrix (delegates to EncodeAdvanced)
  per contract update; EncodePng(m, modulePx = 4, quiet = 4) defaults.
- Tests/dumper adjusted; 17/17 NUnit + node qr-decode gate still green;
  Unity batchmode re-verified clean.

+++ tests/qr-dump/Program.cs
@@ -121,8 +121,8 @@ internal static class Program
         {
             { "name", name },
             { "text", text },
-            { "ecc", (int)m.Ecc },
-            { "level", Levels[(int)m.Ecc] },
+            { "ecc", m.Ecc },
+            { "level", Levels[m.Ecc] },
             { "version", m.Version },
             { "mask", m.Mask },
             { "mode", m.Mode },
